feat(FE): split into notifiers, pages, models
