+++ frontend/Flup/lib/models/Expense.dart
@@ -1,0 +1,8 @@
+class Expense {
+  int price = 0;
+  String title = "";
+  final currency = "CHF";
+  final person = "me";
+
+  Expense(this.price, this.title);
+}
